Add header and the five main sections of webpage
I meant to commit at each section and for some reason I can't patch this but the header contains h1, nav, and image. Each nav item is wrapped in anchor tags. Each section is wrapped in a section tag and consists of an img element, an h2 element, and a p element. The exception is the contact page which has an address tag for the contact info, and the phone and email address are wrapped in an anchor tag